+++ src/js/helpers.js
@@ -18,7 +18,7 @@ import {
   getAllGridRelatedClasses,
 } from './utils'
 import events from './events'
-import { config, defaultTimeout, styles } from './config'
+import { instanceConfig, defaultTimeout, styles } from './config'
 import control from './control'
 import storageAvailable from 'storage-available'
 
@@ -35,6 +35,7 @@ export default class Helpers {
   constructor(formId, layout, formBuilder) {
     this.data = instanceData[formId]
     this.d = instanceDom[formId]
+    this.config = instanceConfig[formId]
     this.doCancel = false
     this.layout = layout
     this.handleKeyDown = this.handleKeyDown.bind(this)
@@ -82,7 +83,7 @@ export default class Helpers {
    */
   beforeStop(event, ui) {
     const _this = this
-    const opts = config.opts
+    const opts = this.config.opts
     const form = _this.d.stage
     const lastIndex = form.childNodes.length - 1
     const cancelArray = []
@@ -196,6 +197,7 @@ export default class Helpers {
     const formData = []
     const d = this.d
     const _this = this
+    const config = this.config
 
     if (form.childNodes.length !== 0) {
       const fields = []
@@ -302,7 +304,7 @@ export default class Helpers {
     const data = this.data
 
     if (!formData) {
-      formData = config.opts.formData
+      formData = this.config.opts.formData
     }
 
     if (!formData) {
@@ -319,7 +321,7 @@ export default class Helpers {
       },
     }
 
-    data.formData = setData[config.opts.dataType](formData) || []
+    data.formData = setData[this.config.opts.dataType](formData) || []
 
     return data.formData
   }
@@ -339,7 +341,7 @@ export default class Helpers {
     }
 
     // save action for current `dataType`
-    data.formData = doSave[config.opts.dataType](minify)
+    data.formData = doSave[this.config.opts.dataType](minify)
 
     // trigger formSaved event
     document.dispatchEvent(events.formSaved)
@@ -365,6 +367,7 @@ export default class Helpers {
    * @return {Object} fieldData
    */
   getAttrVals(field) {
+    const config = this.config
     const fieldData = Object.create(null)
     const attrs = field.querySelectorAll('[class*="fld-"]')
     forEach(attrs, index => {
@@ -651,6 +654,7 @@ export default class Helpers {
    */
   confirmRemoveAll(e) {
     const _this = this
+    const config = this.config
     const formID = e.target.id.match(/frmb-\d{13}/)[0]
     const stage = document.getElementById(formID)
     const i18n = mi18n.current
@@ -683,7 +687,7 @@ export default class Helpers {
 
   addDefaultFields() {
     // Load default fields if none are set
-    config.opts.defaultFields.forEach(field => this.formBuilder.prepFieldVars(field))
+    this.config.opts.defaultFields.forEach(field => this.formBuilder.prepFieldVars(field))
     this.d.stage.classList.remove('empty')
   }
 
@@ -694,7 +698,7 @@ export default class Helpers {
    */
   removeAllFields(stage) {
     const i18n = mi18n.current
-    const opts = config.opts
+    const opts = this.config.opts
     const fields = stage.querySelectorAll(this.formBuilder.fieldSelector)
     const markEmptyArray = []
 
@@ -741,7 +745,7 @@ export default class Helpers {
    * @return {Array|false} fieldOrder
    */
   setFieldOrder($cbUL) {
-    if (!config.opts.sortableControls) {
+    if (!this.config.opts.sortableControls) {
       return false
     }
     const JSON = window.JSON
@@ -839,7 +843,7 @@ export default class Helpers {
 
     this.removeContainerProtection(rowContainer.attr('id'))
 
-    config.opts.onCloseFieldEdit($editPanel[0])
+    this.config.opts.onCloseFieldEdit($editPanel[0])
     document.dispatchEvent(events.fieldEditClosed)
 
     const prevHolder = liContainer.find('.prev-holder')
@@ -909,7 +913,7 @@ export default class Helpers {
     liContainer.insertAfter(rowWrapper)
 
     this.formBuilder.currentEditPanel = $editPanel[0]
-    config.opts.onOpenFieldEdit($editPanel[0])
+    this.config.opts.onOpenFieldEdit($editPanel[0])
     document.dispatchEvent(events.fieldEditOpened)
 
     $(document).trigger('fieldOpened', [{ rowWrapperID: rowWrapper.attr('id') }])
@@ -939,10 +943,10 @@ export default class Helpers {
    * Open a dialog with the form's data
    */
   showData() {
-    const formData = escapeHtml(this.getFormData(config.opts.dataType, true))
+    const formData = escapeHtml(this.getFormData(this.config.opts.dataType, true))
 
     const code = m('code', formData, {
-      className: `formData-${config.opts.dataType}`,
+      className: `formData-${this.config.opts.dataType}`,
     })
 
     this.dialog(m('pre', code), false, 'data-dialog')
@@ -961,7 +965,7 @@ export default class Helpers {
     const fields = form.getElementsByClassName('form-field')
 
     if (!fields.length) {
-      config.opts.notify.warning('No fields to remove')
+      this.config.opts.notify.warning('No fields to remove')
       return false
     }
 
@@ -969,15 +973,15 @@ export default class Helpers {
       const availableIds = [].slice.call(fields).map(field => {
         return field.id
       })
-      config.opts.notify.warning('fieldID required to remove specific fields.')
-      config.opts.notify.warning('Removing last field since no ID was supplied.')
-      config.opts.notify.warning('Available IDs: ' + availableIds.join(', '))
-      fieldID = form.lastChild.id
+      this.config.opts.notify.warning('fieldID required to remove specific fields.')
+      this.config.opts.notify.warning('Removing last field since no ID was supplied.')
+      this.config.opts.notify.warning('Available IDs: ' + availableIds.join(', '))
+      fieldID = availableIds[availableIds.length-1]
     }
 
     const field = document.getElementById(fieldID)
     if (!field) {
-      config.opts.notify.warning('Field not found')
+      this.config.opts.notify.warning('Field not found')
       return false
     }
 
@@ -995,7 +999,7 @@ export default class Helpers {
       }
     })
 
-    const userEvents = Object.assign({}, config.opts.typeUserEvents['*'], config.opts.typeUserEvents[field.type])
+    const userEvents = Object.assign({}, this.config.opts.typeUserEvents['*'], this.config.opts.typeUserEvents[field.type])
 
     if (userEvents && userEvents.onremove) {
       userEvents.onremove(field)
@@ -1059,7 +1063,7 @@ export default class Helpers {
    * @return {Array} subtypes
    */
   processSubtypes(subtypeOpts) {
-    const disabledSubtypes = config.opts.disabledSubtypes
+    const disabledSubtypes = this.config.opts.disabledSubtypes
     // first register any passed subtype options against the appropriate type control class
     for (const fieldType in subtypeOpts) {
       if (subtypeOpts.hasOwnProperty(fieldType)) {
@@ -1136,7 +1140,7 @@ export default class Helpers {
    * @return {HTMLElement[]} formActions btn-group
    */
   formActionButtons() {
-    const opts = config.opts
+    const opts = this.config.opts
     return opts.actionButtons
       .map(btnData => {
         if (btnData.id && opts.disabledActionButtons.indexOf(btnData.id) === -1) {
@@ -1183,7 +1187,7 @@ export default class Helpers {
         events: {
           click: evt => {
             _this.save()
-            config.opts.onSave(evt, _this.data.formData)
+            _this.config.opts.onSave(evt, _this.data.formData)
           },
         },
       },
@@ -1196,8 +1200,8 @@ export default class Helpers {
       // html labels are not available using xml dataType
       opts.disableHTMLLabels = true
     }
-    config.opts = Object.assign({}, { actionButtons: mergedActionButtons }, { fieldEditContainer }, opts)
-    return config.opts
+    _this.config.opts = Object.assign({}, { actionButtons: mergedActionButtons }, { fieldEditContainer }, opts)
+    return _this.config.opts
   }
 
   /**